requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
bertopic==0.16.4
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
db-dtypes==1.3.0
debugpy==1.8.7
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.10.0
gensim==4.3.3
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.21.0
google-auth==2.35.0
google-cloud==0.34.0
google-cloud-bigquery==3.26.0
google-cloud-core==2.4.1
google-cloud-secret-manager==2.22.1
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.14.0
grpcio==1.67.0
grpcio-status==1.67.0
h11==0.14.0
hdbscan==0.8.40
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.18.1
ipywidgets==8.1.5
jedi==0.19.1
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
langdetect==1.0.9
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.21.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
openai==1.52.0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
plotly==5.24.1
prompt_toolkit==3.0.48
proto-plus==1.24.0
protobuf==5.28.2
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pynndescent==0.5.13
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
sentence-transformers==3.3.1
six==1.16.0
smart-open==7.0.5
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
streamlit==1.41.1
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.20.3
toml==0.10.2
torch==2.5.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.46.3
typing_extensions==4.12.2
tzdata==2024.2
umap-learn==0.5.7
urllib3==2.2.3
wcwidth==0.2.13
widgetsnbextension==4.0.13
wrapt==1.16.0
zipp==3.20.2
